requirements.txt
pywin32
pysmb
openpyxl
tqdm